Fix model number of ELEGRP DTR10 Dimmer
When this was originally submitted I used the model number that Amazon listed, but the listing appears to be a copy/paste error, and the DPR10 is actually the non-touch version. This updates to the correct model number shown on the product's box (DTR10) and also notes that this is the single-pole-only version and not the 3-way-compatible version.