Add FL-S122B Smart Switch
This is a smart switch which can be found on AliExpress:
https://www.aliexpress.us/item/3256805310487445.html

The back of the switch contains no interesting markings with respect
to a model/manufacturer (it only says "Smart Switch").  The circuit
board on the inside says FL-S122B-V1.2, which is where I'm getting
this model number from.

The module appears to be a knockoff Tuya CBU-compatible module with
markings "FL_M101_V2" (I suspect manufactured by the same factory).